+++ test-app/tests/jasmine/server/integration/VelocitySpec.js
@@ -1,0 +1,9 @@
+Jasmine.onTest(function () {
+  'use strict';
+
+  describe('Velocity', function () {
+    it('is available via Package["velocity:core"].Velocity', function () {
+      expect(Package['velocity:core'].Velocity).toBeDefined();
+    });
+  });
+});